refactor: remove trait MultiLevelMap, just impl leveled logic for Level
