+++ src/meta/raft-store/src/sm_v002/leveled_store/level.rs
@@ -12,12 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::borrow::Borrow;
 use std::fmt;
+use std::ops::RangeBounds;
 use std::sync::Arc;
 
+use common_meta_types::KVMeta;
+use futures_util::stream::BoxStream;
+use stream_more::KMerge;
+use stream_more::StreamMore;
+
 use crate::sm_v002::leveled_store::level_data::LevelData;
-use crate::sm_v002::leveled_store::leveled_map::MultiLevelMap;
 use crate::sm_v002::leveled_store::map_api::MapApi;
+use crate::sm_v002::leveled_store::map_api::MapApiRO;
+use crate::sm_v002::marked::Marked;
 
 /// One level of state machine data.
 ///
@@ -30,31 +38,17 @@ pub struct Level {
     base: Option<Arc<Level>>,
 }
 
-impl<K> MultiLevelMap<K> for Level
-where
-    K: Ord + fmt::Debug + Send + Sync + 'static,
-    LevelData: MapApi<K>,
-{
-    type API = LevelData;
+impl Level {
+    pub(crate) fn new(data: LevelData, base: Option<Arc<Self>>) -> Self {
+        Self { data, base }
+    }
 
-    fn data<'a>(&'a self) -> &Self::API
-    where Self::API: 'a {
+    pub(crate) fn data(&self) -> &LevelData {
         &self.data
     }
 
-    fn data_mut<'a>(&'a mut self) -> &mut Self::API
-    where Self::API: 'a {
-        &mut self.data
-    }
-
-    fn base(&self) -> Option<&Self> {
+    pub(crate) fn base(&self) -> Option<&Self> {
         self.base.as_ref().map(|x| x.as_ref())
-    }
-}
-
-impl Level {
-    pub fn new(data: LevelData, base: Option<Arc<Self>>) -> Self {
-        Self { data, base }
     }
 
     pub fn new_level(&mut self) {
@@ -86,5 +80,99 @@ impl Level {
 
     pub fn snapshot(&self) -> Option<Arc<Self>> {
         self.base.clone()
+    }
+}
+
+#[async_trait::async_trait]
+impl<K> MapApiRO<K> for Level
+where
+    K: Ord + fmt::Debug + Send + Sync + Unpin + 'static,
+    LevelData: MapApiRO<K>,
+{
+    type V = <LevelData as MapApiRO<K>>::V;
+
+    async fn get<Q>(&self, key: &Q) -> Marked<Self::V>
+    where
+        K: Borrow<Q>,
+        Q: Ord + Send + Sync + ?Sized,
+    {
+        let api = self.data();
+        let got = api.get(key).await;
+
+        if got.is_not_found() {
+            if let Some(base) = self.base() {
+                return base.get(key).await;
+            }
+        }
+        got
+    }
+
+    async fn range<'a, T: ?Sized, R>(&'a self, range: R) -> BoxStream<'a, (K, Marked<Self::V>)>
+    where
+        K: 'a,
+        K: Borrow<T> + Clone,
+        Self::V: Unpin,
+        T: Ord,
+        R: RangeBounds<T> + Clone + Send + Sync,
+    {
+        let a = self.data().range(range.clone()).await;
+
+        let km = KMerge::by(|a: &(K, Marked<Self::V>), b: &(K, Marked<Self::V>)| {
+            let (k1, v1) = a;
+            let (k2, v2) = b;
+
+            assert_ne!((k1, v1.internal_seq()), (k2, v2.internal_seq()));
+
+            // Put entries with the same key together, smaller internal-seq first
+            // Tombstone is always greater.
+            (k1, v1.internal_seq()) <= (k2, v2.internal_seq())
+        })
+        .merge(a);
+
+        let km = if let Some(base) = self.base() {
+            let b = base.range(range).await;
+            km.merge(b)
+        } else {
+            km
+        };
+
+        // Merge entries with the same key, keep the one with larger internal-seq
+        let m = km.coalesce(|(k1, v1), (k2, v2)| {
+            if k1 == k2 {
+                Ok((k1, Marked::max(v1, v2)))
+            } else {
+                Err(((k1, v1), (k2, v2)))
+            }
+        });
+
+        Box::pin(m)
+    }
+}
+
+#[async_trait::async_trait]
+impl<K> MapApi<K> for Level
+where
+    K: Ord + fmt::Debug + Send + Sync + Unpin + 'static,
+    LevelData: MapApi<K>,
+{
+    async fn set(
+        &mut self,
+        key: K,
+        value: Option<(Self::V, Option<KVMeta>)>,
+    ) -> (Marked<Self::V>, Marked<Self::V>)
+    where
+        K: Ord,
+    {
+        // Get from this level or the base level.
+        let prev = self.get(&key).await.clone();
+
+        // No such entry at all, no need to create a tombstone for delete
+        if prev.is_not_found() && value.is_none() {
+            return (prev, Marked::new_tomb_stone(0));
+        }
+
+        // The data is a single level map and the returned `_prev` is only from that level.
+        let (_prev, inserted) = self.data_mut().set(key, value).await;
+        (prev, inserted)
     }
 }
